requirements.txt
# # Core dependencies
# torch>=2.0.0
# transformers>=4.30.0
# sentence-transformers>=2.2.2
# gradio>=3.50.0
# numpy>=1.24.0
# tqdm>=4.65.0
# psutil>=5.9.0

# # PDF processing
# pdfplumber>=0.10.0
# glob2>=0.7

# # AI/ML dependencies
# langchain>=0.0.300
# langchain-community>=0.0.10
# langchain-openai>=0.0.2
# openai>=1.0.0
# chromadb>=0.4.0
# tiktoken>=0.5.0

# # Google Gemini (for additional help)
# google-generativeai>=0.3.0

# # GPU support (optional, only if you have a CUDA GPU)
# cuda-python>=12.0.0
# nvidia-cublas-cu12>=12.0.0
# nvidia-cuda-cupti-cu12>=12.0.0
# nvidia-cuda-nvrtc-cu12>=12.0.0
# nvidia-cuda-runtime-cu12>=12.0.0
# nvidia-cudnn-cu12>=8.9.0
# nvidia-cufft-cu12>=11.0.0
# nvidia-curand-cu12>=10.3.0
# nvidia-cusolver-cu12>=11.4.0
# nvidia-cusparse-cu12>=12.0.0
# nvidia-nccl-cu12>=2.18.0
# nvidia-nvtx-cu12>=12.0.0
# bitsandbytes>=0.45.5
# faiss_cpu>=1.11.0
# accelerate>=1.7.0

# # Vector search
# faiss-cpu

# # TTS (Text-to-Speech)
# pyttsx3<=2.98
# TTS

# # NLP
# spacy
# python-dotenv

# # For spacy model
# en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# # TTS (optional, if you use nari-labs/dia)
# git+https://github.com/nari-labs/dia.git@main
# Fundamental ML/NLP
torch==2.3.0
transformers==4.41.2
sentence-transformers==2.7.0
gradio==4.36.1
numpy
psutil==5.9.0

# PDF processing
pdfplumber==0.10.0
pdfminer.six==20221105

# Google Gemini
google-generativeai==0.3.0

# Vector search
faiss-cpu==1.7.4

# TTS
pyttsx3==2.98
TTS==0.22.0 

# NLP
spacy==3.8.7 # Updated to a version that supports newer typer
python-dotenv==1.0.0 

# For spacy model
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl # <--- UPDATED MODEL URL

# New additions for text processing/
regex==2023.12.25 